make it lossy again oops

+++ src/main/java/com/dynview/Utils/TickTimeHandler.java
@@ -34,7 +34,7 @@ public class TickTimeHandler
 
         if (tickTimer % 20 == 0)
         {
-            meanTickSum += (int) (average(server.getTickTimesNanos()) * 1.0E-6D);
+            meanTickSum += average(server.getTickTimesNanos()) * 1.0E-6D;
             meanTickCount++;
 
             if (tickTimer >= serverTickTimerInterval)
